chore(refactor): Save service with its own bus
